Scroll offset changed to Double

+++ app/src/main/java/com/ovidium/comoriod/data/markups/Markup.kt
@@ -11,7 +11,7 @@ data class Markup(
     val selection: String = "",
     val index: Int,
     val length: Int,
-    val scrollOffset: Int = 0,
+    val scrollOffset: Double = 0.0,
     val bgColor: String = "",
     val timestamp: String = "",
     var tags: List<String> = emptyList()
